Added categories and documents endpoints Added pagination and sorting to student endpoint Added DataLoader class responsible for initial seeding of data to db

+++ src/main/java/com/ahmadabbas/filetracking/backend/document/category/Category.java
@@ -18,10 +18,10 @@ import java.util.StringJoiner;
 @IdClass(SubCategoryPK.class)
 public class Category {
     @Id
-    private Integer parentCategoryId;
+    private Long parentCategoryId;
     @Id
     @GeneratedValue
-    private Integer categoryId;
+    private Long categoryId;
     private String name;
 
     @Override
