+++ src/main/java/com/ahmadabbas/filetracking/backend/document/category/SubCategoryPK.java
@@ -1,13 +1,9 @@
 package com.ahmadabbas.filetracking.backend.document.category;
 
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
 import java.io.Serializable;
 
 
 public class SubCategoryPK implements Serializable {
-    private Integer parentCategoryId;
-    private Integer categoryId;
+    private Long parentCategoryId;
+    private Long categoryId;
 }
